+++ src/main/scala/intent/core/expect.scala
@@ -84,11 +84,20 @@ trait ExpectGivens with
   /**
    * Expect that a key-value tuple exists in the given map
    */
-  def [K, V](expect: Expect[Map[K, V]]) toContain (expected: (K, V)*)
+  def [K, V](expect: Expect[Map[K, V]]) toContain (expected: (K, V))
     (given
        eqq: Eq[V],
        fmt: Formatter[V]
-    ): Expectation = new MapContainExpectation(expect, expected)
+    ): Expectation = new MapContainExpectation(expect, Seq(expected))
+
+  /**
+   * Expect that multiple key-value tuple exists in the given map
+   */
+  def [K, V](expect: Expect[Map[K, V]]) toContainAllOf (expected: (K, V)*)
+  (given
+     eqq: Eq[V],
+     fmt: Formatter[V]
+  ): Expectation = new MapContainExpectation(expect, expected)
 
   // Note: Not using IterableOnce here as it matches Option and we don't want that.
   def [T](expect: Expect[Iterable[T]]) toEqual (expected: Iterable[T])