+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * Add `.toContain` for `Map` objects
   See [toContain](docs/matchers.md#.toContain)
 
+* Add `.toContainAllOf` for `Map` objects
+  See [toContain](docs/matchers.md#.toContainAllOf)
+
 ## 0.3.0 (2019-11-03)
 
 * Upgrade to Dotty 0.20.0-RC1